fix Start-AzAksDashboard and update help files
